# app/Models/Supplier.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\SoftDeletes;

class Supplier extends Model
{
    use HasFactory, SoftDeletes;

    protected $fillable = [
        'name',
        'contact_info',
        'address'
    ];

    public function purchases()
    {
        return $this->hasMany(Purchase::class);
    }

    public function ledger()
    {
        return $this->hasMany(SupplierLedger::class);
    }

}