# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agpos Api",
        "description": "Agpos Assesment Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Fetch a paginated list of suppliers with filters by name",
                "operationId": "6207e57dd1dcd1d6e7125fe144607196",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by supplier name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of suppliers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Supplier"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create a new supplier",
                "operationId": "dc5dc7ccb140cd5b81cab6f42b396691",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Supplier"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/suppliers/{id}": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get details of a specific Supplier",
                "operationId": "7fc447855461518ce36bdeb5a44310a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the supplier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/suppliers/{supplier_id}": {
            "put": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Update an existing supplier",
                "operationId": "45fe0cbcbdb368fd53b16e06db1e8c13",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "ID of the supplier to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Supplier"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Delete a supplier",
                "operationId": "7c44a62556ebd50bfd7058087f6a8d8f",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "ID of the supplier to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supplier deleted successfully"
                    }
                }
            }
        },
        "/api/ledger/payment": {
            "post": {
                "tags": [
                    "Supplier Ledger"
                ],
                "summary": "Add ledger entry for a payment",
                "operationId": "b5ad892328d6570c951d3ca5ad669ad8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_id"
                                ],
                                "properties": {
                                    "payment_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ledger entry added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierLedger"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ledger/{supplier_id}": {
            "get": {
                "tags": [
                    "Supplier Ledger"
                ],
                "summary": "Get supplier ledger",
                "operationId": "e2d1c098f28e3648ed02464cef80eb4c",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "ID of the supplier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier's transaction ledger",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SupplierLedger"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Supplier": {
                "required": [
                    "name",
                    "contact_info",
                    "address"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "contact_info": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "SupplierLedger": {
                "required": [
                    "ledger_id",
                    "supplier_id",
                    "transaction_date",
                    "debit",
                    "credit",
                    "balance",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "ledger_id": {
                        "type": "integer"
                    },
                    "supplier_id": {
                        "type": "integer"
                    },
                    "transaction_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "debit": {
                        "type": "number",
                        "format": "float"
                    },
                    "credit": {
                        "type": "number",
                        "format": "float"
                    },
                    "balance": {
                        "type": "number",
                        "format": "float"
                    },
                    "remarks": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}